Added hardened_malloc init script
Either live-boot or live-config causes the boot/poweroff process
to take a very long time when hardened_malloc is enabled in the
iso. This is not a hardened_malloc issue, as it does not have that
effect on a full install.

Either way, creating a small init script to enable hardened_malloc
at boot once everything else has been initialized fixes the issue.