added dummies
Programs in the base system won't be distributed as debs, but in many
cases, apt needs to be aware of their presence so that their deb
counterpart doesn't get installed at any point.

In the case of man-db, this dummy exists to trick apt into thinking it's
installed so that packages that have it as a hard dependency won't
install it and mess up mandoc.