Third time's the charm
Hopefully, this will permanently fix the issue where the 'cp -r' command
copies the contents of src, rather than the directory itself. I have no
idea why it happens, and in fact, it also happens to the x11 directory
if it gets copied first. Either way, this for loop seems to fix the
problem.