Fix surf displaying blank pages after webkit upd
Refer to https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1054150 for
details.

I am not willing to add more permissions in the apparmor profile in
order to get it to work without passing the variable manually. On top of
that, running surf without WEBKIT_DISABLE_COMPOSITING_MODE=1 makes it
use too much resource anyway. Webkit sucks, we have to deal with it.