handle bsd-mailx and dma
Normally, these wouldn't even be included in
package-lists/base.list.chroot but cron recommends a default MTA and
would install exim instead. So keeping them included and just
overwriting them with the compiled source is a better solution. Of
course that means their files in /var/lib/dpkg/info need to be removed
before the dummies are installed to not mess things up.