Return to chrootcmd instead of arch-chroot
arch-chroot fails to run emerge commands. This isn't an issue with
arch-chroot or emerge, but rather how they behave when paired together.
This also wouldn't happen when chrooted using the arch-chroot command,
but the whole point is to not chroot permanently and chroot just to run
single commands. So there you go.